# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RuiQi 应用防火墙管理系统 API",
        "title": "RuiQi-WAF API",
        "termsOfService": "https://github.com/HUAHUAI23/RuiQi",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/HUAHUAI23/RuiQi",
            "email": "huahua1319873800@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:2333",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前系统配置信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "获取系统配置",
                "responses": {
                    "200": {
                        "description": "获取配置成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ConfigResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "配置不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "使用补丁方式更新系统配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置管理"
                ],
                "summary": "更新系统配置",
                "parameters": [
                    {
                        "description": "配置更新信息",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConfigPatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "配置更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ConfigResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "404": {
                        "description": "配置不存在",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AppConfigDTO": {
            "type": "object",
            "properties": {
                "directives": {
                    "description": "指令配置",
                    "type": "string"
                },
                "logFile": {
                    "description": "日志文件",
                    "type": "string"
                },
                "logFormat": {
                    "description": "日志格式",
                    "type": "string"
                },
                "logLevel": {
                    "description": "日志级别",
                    "type": "string"
                },
                "name": {
                    "description": "应用名称",
                    "type": "string"
                },
                "transactionTTL": {
                    "description": "事务超时时间(毫秒)",
                    "type": "integer",
                    "example": 60000
                }
            }
        },
        "dto.AppConfigPatchDTO": {
            "type": "object",
            "properties": {
                "directives": {
                    "description": "指令配置",
                    "type": "string"
                },
                "logFile": {
                    "description": "日志文件",
                    "type": "string",
                    "example": "/dev/stdout"
                },
                "logFormat": {
                    "description": "日志格式",
                    "type": "string",
                    "example": "console"
                },
                "logLevel": {
                    "description": "日志级别",
                    "type": "string",
                    "example": "info"
                },
                "name": {
                    "description": "应用名称",
                    "type": "string",
                    "example": "coraza"
                },
                "transactionTTL": {
                    "description": "事务超时时间(毫秒)",
                    "type": "integer",
                    "example": 60000
                }
            }
        },
...
        "dto.ConfigPatchRequest": {
            "description": "用于部分更新配置的请求参数",
            "type": "object",
            "properties": {
                "engine": {
                    "description": "引擎配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.EnginePatchDTO"
                        }
                    ]
                },
                "haproxy": {
                    "description": "HAProxy配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.HaproxyPatchDTO"
                        }
                    ]
                },
                "isDebug": {
                    "description": "是否开启调试模式",
                    "type": "boolean",
                    "example": false
                },
                "isK8s": {
                    "description": "是否在K8s环境",
                    "type": "boolean",
                    "example": false
                },
                "isResponseCheck": {
                    "description": "是否检查响应",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "配置名称",
                    "type": "string",
                    "example": "AppConfig"
                }
            }
        },
        "dto.ConfigResponse": {
            "description": "配置响应",
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "engine": {
                    "description": "引擎配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.EngineDTO"
                        }
                    ]
                },
                "haproxy": {
                    "description": "HAProxy配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.HaproxyDTO"
                        }
                    ]
                },
                "id": {
                    "description": "配置ID",
                    "type": "string"
                },
                "isDebug": {
                    "description": "是否开启调试模式",
                    "type": "boolean"
                },
                "isK8s": {
                    "description": "是否在K8s环境",
                    "type": "boolean"
                },
                "isResponseCheck": {
                    "description": "是否检查响应",
                    "type": "boolean"
                },
                "name": {
                    "description": "配置名称",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "dto.EngineDTO": {
            "type": "object",
            "properties": {
                "appConfig": {
                    "description": "应用配置列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AppConfigDTO"
                    }
                },
                "asnDBPath": {
                    "description": "ASN数据库路径",
                    "type": "string"
                },
                "bind": {
                    "description": "引擎绑定地址",
                    "type": "string"
                },
                "cityDBPath": {
                    "description": "城市数据库路径",
                    "type": "string"
                },
                "flowController": {
                    "description": "流量控制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.FlowControllerDTO"
                        }
                    ]
                },
                "useBuiltinRules": {
                    "description": "是否使用内置规则",
                    "type": "boolean"
                }
            }
        },
        "dto.EnginePatchDTO": {
            "type": "object",
            "properties": {
                "appConfig": {
                    "description": "应用配置列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AppConfigPatchDTO"
                    }
                },
                "asnDBPath": {
                    "description": "ASN数据库路径",
                    "type": "string",
                    "example": "/opt/geoip/GeoLite2-ASN.mmdb"
                },
                "bind": {
                    "description": "引擎绑定地址",
                    "type": "string",
                    "example": "127.0.0.1:2342"
                },
                "cityDBPath": {
                    "description": "城市数据库路径",
                    "type": "string",
                    "example": "/opt/geoip/GeoLite2-City.mmdb"
                },
                "flowController": {
                    "description": "流量控制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.FlowControllerPatchDTO"
                        }
                    ]
                },
                "useBuiltinRules": {
                    "description": "是否使用内置规则",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.FlowControllerDTO": {
            "type": "object",
            "properties": {
                "attackLimit": {
                    "description": "攻击频率限制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.LimitConfigDTO"
                        }
                    ]
                },
                "errorLimit": {
                    "description": "错误频率限制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.LimitConfigDTO"
                        }
                    ]
                },
                "visitLimit": {
                    "description": "访问频率限制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.LimitConfigDTO"
                        }
                    ]
                }
            }
        },
        "dto.FlowControllerPatchDTO": {
            "type": "object",
            "properties": {
                "attackLimit": {
                    "description": "攻击频率限制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.LimitConfigPatchDTO"
                        }
                    ]
                },
                "errorLimit": {
                    "description": "错误频率限制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.LimitConfigPatchDTO"
                        }
                    ]
                },
                "visitLimit": {
                    "description": "访问频率限制配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.LimitConfigPatchDTO"
                        }
                    ]
                }
            }
        },
...
        "dto.HaproxyDTO": {
            "type": "object",
            "properties": {
                "backupsNumber": {
                    "description": "备份数量",
                    "type": "integer"
                },
                "configBaseDir": {
                    "description": "配置文件根目录",
                    "type": "string"
                },
                "haproxyBin": {
                    "description": "HAProxy二进制文件路径",
                    "type": "string"
                },
                "spoeAgentAddr": {
                    "description": "SPOE代理地址",
                    "type": "string"
                },
                "spoeAgentPort": {
                    "description": "SPOE代理端口",
                    "type": "integer"
                },
                "thread": {
                    "description": "线程数",
                    "type": "integer"
                }
            }
        },
        "dto.HaproxyPatchDTO": {
            "type": "object",
            "properties": {
                "backupsNumber": {
                    "description": "备份数量",
                    "type": "integer",
                    "example": 5
                },
                "configBaseDir": {
                    "description": "配置文件根目录",
                    "type": "string",
                    "example": "/RuiQi"
                },
                "haproxyBin": {
                    "description": "HAProxy二进制文件路径",
                    "type": "string",
                    "example": "haproxy"
                },
                "spoeAgentAddr": {
                    "description": "SPOE代理地址",
                    "type": "string",
                    "example": "127.0.0.1"
                },
                "spoeAgentPort": {
                    "description": "SPOE代理端口",
                    "type": "integer",
                    "example": 2342
                },
                "thread": {
                    "description": "线程数",
                    "type": "integer",
                    "maximum": 256,
                    "minimum": 0,
                    "example": 4
                }
            }
        },
...
        "dto.LimitConfigDTO": {
            "type": "object",
            "properties": {
                "blockDuration": {
                    "description": "封禁时长（秒）",
                    "type": "integer"
                },
                "burstCount": {
                    "description": "允许的突发请求数",
                    "type": "integer"
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "paramsCapacity": {
                    "description": "缓存容量",
                    "type": "integer"
                },
                "statDuration": {
                    "description": "统计时间窗口（秒）",
                    "type": "integer"
                },
                "threshold": {
                    "description": "阈值",
                    "type": "integer"
                }
            }
        },
        "dto.LimitConfigPatchDTO": {
            "type": "object",
            "properties": {
                "blockDuration": {
                    "description": "封禁时长（秒）",
                    "type": "integer",
                    "example": 600
                },
                "burstCount": {
                    "description": "允许的突发请求数",
                    "type": "integer",
                    "example": 10
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean",
                    "example": true
                },
                "paramsCapacity": {
                    "description": "缓存容量",
                    "type": "integer",
                    "example": 10000
                },
                "statDuration": {
                    "description": "统计时间窗口（秒）",
                    "type": "integer",
                    "example": 60
                },
                "threshold": {
                    "description": "阈值",
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "model.ErrResponse": {
            "description": "错误的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "参数错误"
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.ErrResponseDontShowError": {
            "description": "错误的API响应标准格式,不展示 error",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
        "model.SuccessResponse": {
            "description": "成功的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer {token} 格式进行身份验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}